+++ src/api/v1/apps/orders/services.py
@@ -213,7 +213,8 @@ class OrderService:
                 status_code=status.HTTP_400_BAD_REQUEST,
                 detail="Cannot add items to a closed and checked-out order."
             )
-
+        # let's debug the item_in
+        print(f"{'*' * 30} ITEM_IN IN SERVICE:\n{item_in}")
         # Clean & Direct: Always instantiate a brand-new row for the database
         db_item = OrderDetail(
             order_id=order_id,
@@ -226,7 +227,8 @@ class OrderService:
             extra_charge=item_in.extra_charge,
             prep_status=ItemPrepStatus.QUEUED  # Explicit default configuration tracking
         )
-        
+        # finally debug the final object
+        print(f"{'+' * 30} DB_ITEM IN SERVICE:\n{db_item}")
         self.session.add(db_item)
         is_new_item = True  # Flag marked as True permanently because a row is always added